rename vulkan layer
