+++ source/Core/Base.hpp
@@ -1,6 +1,5 @@
 #pragma once
 
-#include <vulkan/vulkan.h>
 #include <cstdint>
 #include <glm/glm.hpp>
